time_t changes for consideration
plus updates to 2 core files and 11 example programs that use time_t

+++ examples/Time/CosaAlarm/CosaAlarm.ino
@@ -31,14 +31,14 @@
 
 class TraceAlarm : public Alarm {
 public:
-  TraceAlarm(uint8_t id, uint32_t period);
+  TraceAlarm(uint8_t id, uint16_t period);
   virtual void run();
 private:
   uint8_t m_id;
   uint16_t m_tick;
 };
 
-TraceAlarm::TraceAlarm(uint8_t id, uint32_t period) : 
+TraceAlarm::TraceAlarm(uint8_t id, uint16_t period) : 
   Alarm(period), 
   m_id(id), 
   m_tick(0)
